Rename product search to category search
