+++ src/components/CategorySearchScreen/Filters/FilterAppliedChips.tsx
