+++ src/components/CategorySearchScreen/Filters/FilterDiscounts.tsx
