+++ src/components/CategorySearchScreen/Filters/FilterPrice.tsx
