+++ src/components/CategorySearchScreen/Rating.tsx
